yolo-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
onnxruntime==1.16.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
aiofiles==23.2.0